Documents/TSED/Proyecto_biblioteca_sdp/conexion/conexion.py
import mysql.connector

def get_connection():
    return mysql.connector.connect(
        host='localhost',
        user='root',
        password='',          # Cambia por tu contraseña MySQL
        database='biblioteca_db'
    )
