Documents/TSED/Proyecto_biblioteca_sdp/requirements.txt
flask
flask-login
mysql-connector-python
reportlab
